+++ e2e/helpers/gql.ts
@@ -1,0 +1,5 @@
+// TODO: refactor
+export const gql = (q: string, variables?: Record<string, unknown>) => ({
+  query: q,
+  variables,
+})